fastapi_mini_project/requirements.txt
aerich==0.9.2
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
asyncclick==8.3.0.7
asyncpg==0.31.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
build==1.4.0
CacheControl==0.14.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cleo==2.1.0
click==8.3.1
comm==0.2.3
crashtest==0.4.1
cryptography==46.0.5
debugpy==1.8.19
decorator==5.2.1
dictdiffer==0.9.0
distlib==0.4.0
dnspython==2.8.0
dulwich==0.21.7
ecdsa==0.19.1
email-validator==2.3.0
executing==2.2.1
fastapi==0.135.1
fastjsonschema==2.21.2
filelock==3.20.3
Flask==3.1.2
flask-sock==0.7.0
h11==0.16.0
idna==3.11
installer==0.7.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
iso8601==2.1.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.7.0
jupyter_core==5.9.1
keyring==24.3.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
more-itertools==10.8.0
msgpack==1.1.2
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.5
passlib==1.7.4
pexpect==4.9.0
pkginfo==1.12.1.2
platformdirs==4.5.1
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
prompt_toolkit==3.0.52
psutil==7.2.1
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pypika-tortoise==0.6.3
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
pytz==2026.1.post1
pyzmq==27.1.0
RapidFuzz==3.14.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
stack-data==0.6.3
starlette==0.52.1
tomlkit==0.14.0
tornado==6.5.4
tortoise-orm==0.25.4
traitlets==5.14.3
trove-classifiers==2026.1.14.14
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
virtualenv==20.36.1
wcwidth==0.2.14
Werkzeug==3.1.5
wsproto==1.3.2
xattr==1.3.0